Update re-frame's database model
In order to use the re-frame's subscriptions model properly, updates
database to store the current elements to be drawn.
Now, `:matrix-effect.subs/elements` can be used to suscribe to the
current elements of the frame. They are returned on a sequence with this
form `((c_1 x_1 y_1), (c_2 x_2 y_2), ...)`.