Wrap `canvas` on React class
Although we are currently using re-frame, dividing pure components from
the impure ones, modifying the `canvas` via
```
:matrix-effect.canvas.util/get-canvas ⟶
getElementById ⟶
:matrix-effect.canvas.util/draw-rectangle
```
(using coeffects, effects) not fully sticks with the re-frame's
philosophy. It proposes that the data should be sended to views through
subscriptions, and the new characters to paint on `canvas`, are at the
end of the day, this kind of data.

Creates a wrapper to encapsulate the impurity of `canvas`, which now
receives the new elements to be painted, and internally carry out the
necessary processes to paint them. `:matrix-effect.events/draw-frame`
effect is deprecated and instead `:matrix-effect.events/next-frame` now
just updates the database with the new model.