WIP: Changes the underlying model
A new entity symbol is defined. Each symbol contains:

- The character to be drawn
- The column it corresponds to
- The row where it will be drawn

In this way we can have more encapsulation and operate on a higher
level (i.e. draw a symbol, wherever it may be).

On this same commit, some of the arguments of the canvas modifying
functions are removed (color, char width & height). These never change,
so it is a little easier on the eyes of the caller if we simply hide
them. The `draw-rectangle!` function is marked as private, as it is
pretty low-level. To replace the uses of `draw-rectangle!` two new
public functions are included:

- draw-background!
- fade-canvas!

Hopefully, the names are a little more descriptive of what each call
is acomplishing.

A new function `draw-symbol!` is created, again the
motivation was to hide somethings from the top-level caller (like the
exact position where the char will be drawn).

Finally, some function names in the events namespace have been changed
in order to aid with code comprehension.